Added views and controllers
